test: Update extensions test env
